web/requirements.txt
Flask
flask_restful
pymongo
bcrypt
numpy
keras
tensorflow
requests
pillow